Stops the internal progressbar from restarting
This addresses [issue #1](https://github.com/fsaxen/ParforProgMon/issues/1)  which causes the internal `progressbar` to be restarted and clear the labels. 

The problem was that for fairly long iterations, if no progress was made within the defined update period, the internall `progressbar` would be called with all-zero inputs, which would reset, among else, it removing labels.

The introduced change was to add a small epsilon to stop the progressbar from getting cleared.

TODO: before the first iteration is completed, the estimated time is "Inf week", this should be made more meaningful like "Waiting", "Waiting for the first iteration to complete", "Estimating ETA..." or something similar.